requirements.txt
PyPDF2==3.0.1
langchain==0.1.0
python-dotenv==1.0.0
openai==1.12.0
streamlit==1.31.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2 